+++ src/web/src/editor/CollaborativeEditor.tsx
@@ -18,6 +18,7 @@ import { handleImageDrop, handleImagePaste } from './imageUpload'
 import { setSlashCommandStorage } from './slash/items'
 import { setDynamicBlockStorage } from './dynamicBlock'
 import { DynamicBlockMenu } from './DynamicBlockMenu'
+import { TocMenu } from './TocMenu'
 import type { CollabConnection } from './CollabStatus'
 
 type Props = {
@@ -159,6 +160,7 @@ export function CollaborativeEditor({
       {editor && <DateMenu editor={editor} />}
       {editor && <LayoutMenu editor={editor} />}
       {editor && <DynamicBlockMenu editor={editor} />}
+      {editor && <TocMenu editor={editor} />}
       <EditorContent editor={editor} className="editor__content" />
     </div>
   )